+++ admin/patch/mollie/6_addDateModifiedField.php
@@ -1,0 +1,8 @@
+<?php
+use comercia\Util;
+
+return function () {
+    Util::patch()->table("mollie_payments")
+        ->addField("date_modified", "DATETIME NOT NULL")
+        ->update();
+};