+++ system/comercia/config.php
@@ -6,11 +6,13 @@ class Config
     var $model;
     var $store_id;
     var $data = [];
+    var $db;
 
     function __construct($store_id = 0)
     {
         $this->model = Util::load()->model("setting/setting");
         $this->store_id = $store_id;
+        $this->db = Util::registry()->get("db");
         $data = Util::db()->query("SELECT * FROM " . DB_PREFIX . "setting WHERE store_id = " . $store_id . "");
         foreach ($data as $value) {
             if (!$value['serialized']) {
@@ -57,6 +59,16 @@ class Config
         foreach ($items as $key => $val) {
             $this->data[$key] = $val;
         }
+    }
+
+    function setValue($code, $key, $value = '')
+    {
+        if (!is_array($value)) {
+            $this->db->query("INSERT INTO " . DB_PREFIX . "setting SET store_id = '" . (int)$this->store_id . "', `code` = '" . $this->db->escape($code) . "', `key` = '" . $this->db->escape($key) . "', `value` = '" . $this->db->escape($value) . "'");
+        } else {
+            $this->db->query("INSERT INTO " . DB_PREFIX . "setting SET store_id = '" . (int)$this->store_id . "', `code` = '" . $this->db->escape($code) . "', `key` = '" . $this->db->escape($key) . "', `value` = '" . $this->db->escape(json_encode($value, true)) . "', serialized = '1'");
+        }
+        $this->data[$key] = $value;
     }
 }